pyproject.toml
[project]
name = "pill-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.25.0",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

